requirements-dev.txt
django==2.1.9
django-debug-toolbar==1.11
django_rest_framework_camel_case==2.0.0
djangorestframework==3.9.1
pytest==4.1.1
pytest_factoryboy==2.0.2
pytest_django==3.4.5
flake8==3.6.0
inflection>=0.3.1
